FileSharing_Project/requirements.txt
djangorestframework==3.12
jieba==0.42.1
oss2==2.16.0
PyJWT==2.4.0
PyMySQL==1.0.2
requests==2.27.1
six==1.16.0
django~=2.2
django_redis==4.7.0
pillow~=7.2.0
celery~=4.1.1